Added user screen, domains, created permission to access screens

+++ routes/web.php
@@ -9,33 +9,28 @@ use Illuminate\Support\Facades\Route;
 
     Route::get('/{erro?}', [LoginController::class,'index'])->name('login');
     Route::post('/', [LoginController::class,'authentication'])->name('login');
-    Route::get('/register', [UserController::class,'create'])->name('login.create');
 
 
 //for more information about rotues, use the command: php artisan route:list
-Route::middleware(
-    [
-        'auth',
-        'needsRole:admin'
-    ]
-    )->prefix('/admin')->group(function(){
+Route::middleware(['auth'])->prefix('/system')->group(function(){
     
-    Route::get('/publisher', [PublisherController::class, 'index'])->name('publisher.index');
-    Route::post('/publisher', [PublisherController::class, 'store'])->name('publisher.store');
+    Route::get('/publisher', [PublisherController::class, 'index'])->name('publisher.index')->middleware('needsRole:admin');
+    Route::post('/publisher', [PublisherController::class, 'store'])->name('publisher.store')->middleware('needsRole:admin');
 
-    Route::get('/publisher/create', [PublisherController::class, 'create'])->name('publisher.create');
-    Route::get('/publisher/{publisher}', [PublisherController::class, 'show'])->name('publisher.show');
+    Route::get('/publisher/create', [PublisherController::class, 'create'])->name('publisher.create')->middleware('needsRole:admin');
+    Route::get('/publisher/{publisher}', [PublisherController::class, 'show'])->name('publisher.show')->middleware('needsRole:admin');
 
-    Route::get('/publisher/{publisher}/edit', [PublisherController::class, 'edit'])->name('publisher.edit');
-    Route::put('/publisher/{publisher}', [PublisherController::class, 'update'])->name('publisher.update');
+    Route::get('/publisher/{publisher}/edit', [PublisherController::class, 'edit'])->name('publisher.edit')->middleware('needsRole:admin');
+    Route::put('/publisher/{publisher}', [PublisherController::class, 'update'])->name('publisher.update')->middleware('needsRole:admin');
 
-    Route::delete('/publisher/{publisher}', [PublisherController::class, 'destroy'])->name('publisher.destroy');
+    Route::delete('/publisher/{publisher}', [PublisherController::class, 'destroy'])->name('publisher.destroy')->middleware('needsRole:admin');
 
 
     Route::get('/logout', [LoginController::class,'logout'])->name('logout');
 
 
-    Route::get('/domain', [DomainController::class, 'index'])->name('domain.index');
+    Route::get('/domain', [DomainController::class, 'index'])->name('domain.admin.index')->middleware('needsRole:admin');
+    Route::get('/domain', [DomainController::class, 'listarPorPublisher'])->name('domain.publisher.index')->middleware('needsRole:publisher');
     Route::post('/domain', [DomainController::class, 'store'])->name('domain.store');
 
     Route::get('/domain/create', [DomainController::class, 'create'])->name('domain.create');
@@ -47,48 +42,14 @@ Route::middleware(
     Route::delete('/domain/{domain}', [DomainController::class, 'destroy'])->name('domain.destroy');
 
 
-    Route::get('/user', [UserController::class, 'index'])->name('user.index');
-    Route::post('/user', [UserController::class, 'store'])->name('user.store');
+    Route::get('/user', [UserController::class, 'index'])->name('user.index')->middleware('needsRole:admin');
+    Route::post('/user', [UserController::class, 'store'])->name('user.store')->middleware('needsRole:admin');
 
-    Route::get('/user/create', [UserController::class, 'create'])->name('user.create');
-    Route::get('/user/{user}', [UserController::class, 'show'])->name('user.show');
+    Route::get('/user/create', [UserController::class, 'create'])->name('user.create')->middleware('needsRole:admin');
+    Route::get('/user/{user}', [UserController::class, 'show'])->name('user.show')->middleware('needsRole:admin');
 
-    Route::get('/user/{user}/edit', [UserController::class, 'edit'])->name('user.edit');
-    Route::put('/user/{user}', [UserController::class, 'update'])->name('user.update');
+    Route::get('/user/{user}/edit', [UserController::class, 'edit'])->name('user.edit')->middleware('needsRole:admin');
+    Route::put('/user/{user}', [UserController::class, 'update'])->name('user.update')->middleware('needsRole:admin');
 
-    Route::delete('/user/{user}', [UserController::class, 'destroy'])->name('user.destroy');
-});
-
-Route::middleware(
-    [
-        'auth',
-        'needsRole:publisher'
-    ]
-    )->prefix('/publisher')->group(function(){
-    
-    Route::get('/publisher', [PublisherController::class, 'index'])->name('publisher.index');
-    Route::post('/publisher', [PublisherController::class, 'store'])->name('publisher.store');
-
-    Route::get('/publisher/create', [PublisherController::class, 'create'])->name('publisher.create');
-    Route::get('/publisher/{publisher}', [PublisherController::class, 'show'])->name('publisher.show');
-
-    Route::get('/publisher/{publisher}/edit', [PublisherController::class, 'edit'])->name('publisher.edit');
-    Route::put('/publisher/{publisher}', [PublisherController::class, 'update'])->name('publisher.update');
-
-    Route::delete('/publisher/{publisher}', [PublisherController::class, 'destroy'])->name('publisher.destroy');
-
-
-    Route::get('/logout', [LoginController::class,'logout'])->name('logout');
-
-
-    Route::get('/domain', [DomainController::class, 'index'])->name('domain.index');
-    Route::post('/domain', [DomainController::class, 'store'])->name('domain.store');
-
-    Route::get('/domain/create', [DomainController::class, 'create'])->name('domain.create');
-    Route::get('/domain/{domain}', [DomainController::class, 'show'])->name('domain.show');
-
-    Route::get('/domain/{domain}/edit', [DomainController::class, 'edit'])->name('domain.edit');
-    Route::put('/domain/{domain}', [DomainController::class, 'update'])->name('domain.update');
-
-    Route::delete('/domain/{domain}', [DomainController::class, 'destroy'])->name('domain.destroy');
+    Route::delete('/user/{user}', [UserController::class, 'destroy'])->name('user.destroy')->middleware('needsRole:admin');
 });
